Remove app.py, and move all of the logic out
