--- app.py
@@ -1,23 +1,0 @@
-import os
-
-from flask import Flask
-from flask_login import LoginManager
-from flask_sqlalchemy import SQLAlchemy
-
-app = Flask(__name__)
-app.config.from_object(os.environ['APP_SETTINGS'])
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
-db = SQLAlchemy(app)
-login_manager = LoginManager()
-login_manager.init_app(app)
-
-from models import Users
-
-
-@login_manager.user_loader
-def user_loader(user_email):
-    return Users.query.get(user_email)
-
-
-if __name__ == '__main__':
-    app.run()
