+++ app/__init__.py
@@ -1,0 +1,17 @@
+from flask import Flask
+from flask_login import LoginManager
+from flask_sqlalchemy import SQLAlchemy
+
+from config import config
+
+db = SQLAlchemy()
+login_manager = LoginManager()
+
+
+def create_app(config_name):
+    app = Flask(__name__)
+    app.config.from_object(config[config_name])
+    db.init_app(app)
+    login_manager.init_app(app)
+
+    return app
