+++ config.py
@@ -25,3 +25,10 @@ class LocalConfig(Config):
 
 class TestingConfig(Config):
     TESTING = True
+
+
+config = {
+    'development': DevelopmentConfig,
+    'testing': TestingConfig,
+    'production': ProductionConfig,
+}
